LayoutTests/crypto/subtle/hmac-sign-verify.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test HMAC sign and verify functions.");

jsTestIsAsync = true;

if (!window.subtle)
    window.crypto.subtle = window.crypto.webkitSubtle;

var hmacKey = asciiToArrayBuffer('a');
var extractable = true;

debug("Importing a raw HMAC key from string literal...");
crypto.subtle.importKey("raw", hmacKey, {name: 'hmac', hash: {name: 'sha-1'}}, extractable, ["sign", "verify"]).then(function(result) {
    key = result;
    shouldBe("key.type", "'secret'");
    shouldBe("key.extractable", "true");
    shouldBe("key.algorithm.name", "'hmac'");
    shouldBe("key.algorithm.length", "1"); // See <https://www.w3.org/Bugs/Public/show_bug.cgi?id=23098>.
    shouldBe("key.algorithm.hash.name", "'sha-1'");
    shouldBe("key.usages", "['sign', 'verify']");

    debug("Using the key to sign 'foo'...");
    return crypto.subtle.sign(key.algorithm, key, [asciiToArrayBuffer('foo')]);
}).then(function(result) {
    printAcceptedResult(result);

    debug("Verifying the signature...");
    return crypto.subtle.verify(key.algorithm, key, result, [asciiToArrayBuffer('foo')]);
}).then(function(result) {
    verificationResult = result;
    shouldBe("verificationResult", "true");
    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
