LayoutTests/crypto/subtle/hmac-sign-verify-expected.txt

Test HMAC sign and verify functions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw HMAC key from string literal...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'hmac'
PASS key.algorithm.length is 1
PASS key.algorithm.hash.name is 'sha-1'
PASS key.usages is ['sign', 'verify']
Using the key to sign 'foo'...
    = [be bb c0 2e 46 b0 f8 11 83 f4 0c 25 dc e2 3e 50 45 d6 55 19]
Verifying the signature...
PASS verificationResult is true
PASS successfullyParsed is true

TEST COMPLETE
